+++ installation-bundle/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Contao installation bundle change log
 
+### DEV
+
+ * Check if the Contao framework has been initialized when adding the user agent string (see standard-edition#64).
+
 ### 4.4.1 (2017-07-12)
 
  * Correctly set the "overwriteMeta" field during the database update (see contao/core-bundle#888).